refactor(field-trip): move non-payment fields to group info
