+++ lib/components/admin/editable-section.js
@@ -56,7 +56,12 @@ export default class EditableSection extends Component {
     return (
       <>
         {children}
-        <span className='pull-right' style={{marginRight: '15px'}}>
+        <span
+          className='pull-right'
+          // Limit height so that it does not collide with form elements
+          // below when actively editing.
+          style={{height: '10px', marginRight: '15px'}}
+        >
           {!isEditing
             ? <Button
               bsSize='xsmall'
