+++ lib/components/admin/field-trip-details.js
@@ -18,44 +18,18 @@ import {
   Header,
   InlineHeader,
   P,
-  Text
+  Text,
+  Val
 } from './styled'
 import TripStatus from './trip-status'
 import Updatable from './updatable'
-import {getGroupSize} from '../../util/call-taker'
+import {
+  getGroupSize,
+  GROUP_FIELDS,
+  PAYMENT_FIELDS,
+  TICKET_TYPES
+} from '../../util/call-taker'
 
-const TICKET_TYPES = {
-  own_tickets: 'Will use own tickets',
-  hop_new: 'Will purchase new Hop Card',
-  hop_reload: 'Will reload existing Hop Card'
-}
-const PAYMENT_PREFS = {
-  request_call: 'Call requested at provided phone number',
-  phone_cc: 'Will call in credit card info to TriMet',
-  fax_cc: 'Will fax credit card info to TriMet',
-  mail_check: 'Will mail check to TriMet'
-}
-
-const inputProps = {
-  min: 0,
-  step: 1,
-  type: 'number'
-}
-const GROUP_FIELDS = [
-  {inputProps, fieldName: 'numStudents', label: 'students 7 or older'},
-  {inputProps, fieldName: 'numFreeStudents', label: 'students under 7'},
-  {inputProps, fieldName: 'numChaperones', label: 'chaperones'}
-]
-const PAYMENT_FIELDS = [
-  {label: 'Ticket type', fieldName: 'ticketType', options: TICKET_TYPES},
-  {label: 'Payment preference', fieldName: 'paymentPreference', options: PAYMENT_PREFS},
-  {label: 'Invoice required', fieldName: 'requireInvoice', options: ['Yes', 'No']},
-  {label: 'Class Pass Hop Card #', fieldName: 'classpassId'},
-  {label: 'Credit card type', fieldName: 'ccType'},
-  {label: 'Name on credit card', fieldName: 'ccName'},
-  {label: 'Credit card last 4 digits', fieldName: 'ccLastFour'},
-  {label: 'Check/Money order number', fieldName: 'checkNumber'}
-]
 /**
  * Shows the details for the active Field Trip Request.
  */
@@ -87,10 +61,12 @@ class FieldTripDetails extends Component {
     if (!request) return null
     const {
       id,
+      invoiceRequired,
       notes,
       schoolName,
       submitterNotes,
       teacherName,
+      ticketType,
       travelDate
     } = request
     const {fieldTrip} = callTaker
@@ -160,6 +136,8 @@ class FieldTripDetails extends Component {
           <Half>
             <P><B>{schoolName}</B></P>
             <P>Teacher: {teacherName}</P>
+            <P>Ticket type: <Val>{TICKET_TYPES[ticketType]}</Val></P>
+            <P>Invoice required: <Val>{invoiceRequired ? 'Yes' : 'No'}</Val></P>
             <P>
               <Updatable
                 fieldName='Teacher notes'
