+++ lib/util/call-taker.js
@@ -2,6 +2,40 @@ import {isTransit} from '@opentripplanner/core-utils/lib/itinerary'
 
 import {getRoutingParams} from '../actions/api'
 
+export const TICKET_TYPES = {
+  own_tickets: 'Will use own tickets',
+  hop_new: 'Will purchase new Hop Card',
+  hop_reload: 'Will reload existing Hop Card'
+}
+
+const PAYMENT_PREFS = {
+  request_call: 'Call requested at provided phone number',
+  phone_cc: 'Will call in credit card info to TriMet',
+  fax_cc: 'Will fax credit card info to TriMet',
+  mail_check: 'Will mail check to TriMet'
+}
+
+const positiveIntInputProps = {
+  min: 0,
+  step: 1,
+  type: 'number'
+}
+
+export const GROUP_FIELDS = [
+  {inputProps: positiveIntInputProps, fieldName: 'numStudents', label: 'students 7 or older'},
+  {inputProps: positiveIntInputProps, fieldName: 'numFreeStudents', label: 'students under 7'},
+  {inputProps: positiveIntInputProps, fieldName: 'numChaperones', label: 'chaperones'}
+]
+
+export const PAYMENT_FIELDS = [
+  {label: 'Payment preference', fieldName: 'paymentPreference', options: PAYMENT_PREFS},
+  {label: 'Class Pass Hop Card #', fieldName: 'classpassId'},
+  {label: 'Credit card type', fieldName: 'ccType'},
+  {label: 'Name on credit card', fieldName: 'ccName'},
+  {label: 'Credit card last 4 digits', fieldName: 'ccLastFour'},
+  {label: 'Check/Money order number', fieldName: 'checkNumber'}
+]
+
 function placeToLatLonStr (place) {
   return `${place.lat.toFixed(6)},${place.lon.toFixed(6)}`
 }
